+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * The plugin is now compatible with and requires Chartist v1 API
 * The plugin provides an ESM, CSJ and UMD version
+* The "class" option accepts a single class name or an array of class names
 * Sources were converted to TypeScript
 
 ## [1.0.0] - 2nd January 2023