+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * The plugin is now compatible with and requires Chartist v1 API
 * The plugin provides an ESM, CSJ and UMD version
+* The "class" option accepts a single class name or an array of class names
 * Sources were converted to TypeScript
 
 ----