+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [2.0.1] - 20th September 2026
 
 ### Changed
 - Clean up mixed Chartist imports ([#2](https://github.com/stklcode/chartist-plugin-tooltip/pull/2))
@@ -112,7 +112,8 @@ The `s` at the end of the tooltip is now missing.
 - Removed Ember reference
 
 
-[unreleased]: https://github.com/stklcode/chartist-plugin-tooltip/compare/v2.0.0...HEAD
+[unreleased]: https://github.com/stklcode/chartist-plugin-tooltip/compare/v2.0.1...HEAD
+[2.0.1]: https://github.com/stklcode/chartist-plugin-tooltip/compare/v2.0.0..v2.0.1
 [2.0.0]: https://github.com/stklcode/chartist-plugin-tooltip/compare/v1.0.0..v2.0.0
 [1.0.0]: https://github.com/stklcode/chartist-plugin-tooltip/compare/v0.1.4...v1.0.0
 [0.1.4]: https://github.com/stklcode/chartist-plugin-tooltip/compare/v0.1.3...v0.1.4